notebooks/ML_ShinyDash/App/requirements.txt
shiny==1.0.0
databricks-sdk
pandas
plotly
pyarrow
